feat: v0-72-63m + v0-72-64m — loss cap, strategy bug fixes, kill switch, loss streak
v0-72-63m:
- EGW00201 explict error detection in client.ts (403 body parse)
- 1.5% per-trade loss cap in position-sizer.ts (≤₩45,000 on 3M position)
- seed-capital.ts: paper cap 기반 동적 계산 연동

v0-72-64m — 4 strategy bugs fixed:
- opening-bell-job: kill switch bypass 차단 (isKillSwitchActive 체크 추가)
- crash-profit.ts: CAUTION 상태 이중 헷지 방지 (신규 매수 스킵)
- eod-betting-job: runner.ts 종료예약 스케줄 충돌 해결 + SWING cross-close 차단
- kill-switch.ts: paper/live 분리 자동 리셋 함수 추가 (resetKillSwitchForMode)
- runner.ts: 08:50 KR + 22:20 OVERSEAS 킬스위치 자동 리셋 (paper/live 동시)
- settings.ts: 모드별 킬스위치 API 개선

risk & position:
- risk-engine.ts: 1.5% 손실 상한 적용, 연속손실 배율 연동
- loss-streak.ts: 연속손실 트래커 신규 모듈 (2연속=0.7x, 3+연속=0.5x)

db:
- migration 070: paper 고아 체인 정리 (migration 069 이후 OPEN 체인 불일치 해소)
- models.ts: 필드 추가

dashboard:
- builder.ts: [PAPER_TOTAL_DEBUG] 로그 추가 (총자산 성분 분리 진단)
- holding-check-job.ts: break-even protection 로직 보강

+++ src/api/routes/dashboard/overseas-sell.ts
@@ -53,7 +53,7 @@ export function registerOverseasSellRoutes(app: Hono) {
           }
           await tx.query(
             `INSERT INTO overseas_state (key, value) VALUES ('cash_paper', $1::text)
-             ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $1)::text`,
+             ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $1::numeric)::text`,
             [proceeds]);
           await tx.query(
             `INSERT INTO orders (stock_code, side, order_type, quantity, price, filled_quantity, filled_price, kis_order_no, status, trading_mode, trigger_source, ai_reasoning, avg_buy_price)
@@ -157,7 +157,7 @@ export function registerOverseasSellRoutes(app: Hono) {
         if (isPaper) {
           await tx.query(
             `INSERT INTO overseas_state (key, value) VALUES ($1, $2::text)
-             ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $2)::text`,
+             ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $2::numeric)::text`,
             [cashKey, proceeds]);
         }
         await tx.query(
@@ -233,7 +233,7 @@ export function registerOverseasSellRoutes(app: Hono) {
           if (isPaper) {
             await tx.query(
               `INSERT INTO overseas_state (key, value) VALUES ($1, $2::text)
-               ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $2)::text`,
+               ON CONFLICT (key) DO UPDATE SET value = (CAST(overseas_state.value AS NUMERIC) + $2::numeric)::text`,
               [cashKey, proceeds]);
           }
           await tx.query(
